+++ src/module/Process.hpp
@@ -17,5 +17,6 @@ namespace be {
         static JSValue readMac(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue gc(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue ref(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue getRunningTime(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
     } ;
 }